nntrain can now take uint8 data, visualize has new boolean argument whether to contrast normalize each image. nnvisualize projects each neuron up and down in net and uses visualize to plot each neuron receptive field/ output

+++ NN/nntrain.m
@@ -9,8 +9,8 @@ function net = nntrain(net, x, y, opts)
         kk = randperm(m);
         tic;
         for l=1:numbatches
-            batch_x = x(kk((l-1)*opts.batchsize+1:l*opts.batchsize),:);
-            batch_y = y(kk((l-1)*opts.batchsize+1:l*opts.batchsize),:);
+            batch_x = double(x(kk((l-1)*opts.batchsize+1:l*opts.batchsize),:));
+            batch_y = double(y(kk((l-1)*opts.batchsize+1:l*opts.batchsize),:));
             net = nnff(net, batch_x, batch_y);
     %             if(rand() < 1e-3)
     %                 disp 'Performing numerical gradient checking ...';
